Added various major features (see full commit message):
- Syntax of defcomponent was changed so user can specify optional attributes with default values. This could be useful for storing system calculated data inside the entity attributes itself which are not important to the entity definition.

The syntax was flattened out
(ecs:defcomponent some-component
  (some-attribute1 some-default-value)
  (some-attribute2 some-default-value2)
  some-attribute3 ;; parentheses not strictly necessary if no default value needed
  [...])

- Entity deletion queue
This concerns how an entity can be deleted from a system without affecting iteration.

some other approaches include adding a tag for dead entities and then manually coding in a dead entity remover process of some sort (it technically can't be an entity system)
there is a more general problem though and that is how to perform logic on a collection where the collection can be modified by the iteration. It's actually very complicated. Lots of ECS implementations choose to mitigate the effects of the issue.
take Ashley ECS for example: https://github.com/libgdx/ashley/issues/44 - this actually has the risk of skipping an entity if the current entity is deleted.
Long reddit discussion: https://www.reddit.com/r/gamedev/comments/3xtcwe/learning_entitycomponent_system_deleting_entities/

The deletion queue can be processed after the end of each system's iteration. It isnt perfect but I think it is pretty unlikely that any practical logic involves processing the same entity twice in one system even if it polled from outside system's required componenets (e.g manually looping over all-entites in the body of a system).

- New method for system communication
systems can pass messages by using marks and flags
marks are like tags except they are supported by hash table. Marks can be turned on or off by calling (ecs:mark [entity] :flag-name) or (ecs:demk [entity] :flag-name)
flags are like marks except they are reset to nil before the next iteration of the same system, allowing any system to have a chance to operate on that flag before it turns off on the next frame.
(ecs:flag [entity] :flag-name)

Examples:
(when (ecs:some-marks-p #id :is-flying)
  (do-flying-logic))

(defun do-flying-logic (#id)
  ;;land
  (if (is-on-ground #id)
    (ecs:demk #id :is-flying))) ;; turn off flying

(when ((ecs:some-flags-p #id :is-in-air)
       (ecs:no-marks-p #id :is-flying)) ;; in air but not flying
  (do-gravity-logic))

When flags or marks are turned off, they are simply set to nil in the hash table.
There's some code in there which keeps track of which system set what flag but this is meant for debugging scenarios (entity is flying? why?? oh, it's because x system didn't reset it properly/set it when it's not supposed to be - good luck trying to find that out without field value change breakpoints)

Alternative to these approaches is doing something like storing messages in a queue.
they might feel "more correct" or "more functional" but the downside to that approach is that it can (theoretically) cause a lot of garbage and thrashing when they are consumed.

By placing messages and flags in a hash table and not deleting them immediately, the old entries can be cleared out and garbage collected when appropriate e.g during level change/loading screens then calling (gc/collect) or something instead of random stop-the-world garbage collection while playing. TODO add inbuilt mechanism to do this somewhere maybe.

Yes, the entity deletion queue is still subject to this, but whatever. Just dont create and delete entities every frame, lol.

ISSUES:
im not sure whether abstracting away complicated iteration patterns was actually a good idea or not. It might be one of those "catch-most-but-not-all" abstractions. Could try introducing families instead like Ashley does.