Support for custom iteration systems.
uses the exact same syntax as the regular systems, however,
the bindings of the requirement letargs are a list of all matching entities at once rather than individual entities in iteration.
this leaves the system itself free to specify how the matching entities should be iterated.

this also gets rid of the boilerplate ()s after the letargs i was really annoyed about
as per git diff, it hit me like a pile of bricks when i realised i could just change what the bindings meant to get this.